fix: remove unchecked header access in UA hook
This change updates `CustomUserAgentHook` to fix an exception that was
happening in Chrome. The `User-Agent` header cannot be customised in
that browser so the value of `request.headers.get("user-agent")` was
null. However, the code was using a type assertion that did not consider
this.

To fix, we now get the SDK version from `SDK_METADATA` object available
with the SDK and we set an alternative header if setting `User-Agent`
silently failed.

More info: https://issues.chromium.org/issues/40450316